Users, login and backend tests working - Tasks 4.1-4.22 done

+++ controllers/blogs.js
@@ -1,25 +1,44 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const blog = require('../models/blog')
-//const {result} = require('lodash')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 //-------------------------------
-//----------- GET ALL -----------
+//-------- GET ALL BLOGS --------
 //-------------------------------
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs)
+    const blogs = await Blog
+        .find({})
+        .populate('user', {username: 1, name: 1})
+    response.json(blogs.map(b => b.toJSON()))
 })
 
 //-------------------------------
-//---------- POST ONE -----------
+//-------- POST ONE BLOG --------
 //-------------------------------
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', async (request, response, next) =>
+{
+    
     const blog = new Blog(request.body)
+    if (!request.token)
+    {
+        return response.status(401).json({error: 'Token required'})
+    }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id)
+    {
+        return response.status(401).json({error: 'Authorization failed'})
+    }
+    
+    
+    const user = await User.findById(decodedToken.id)
+    blog.user = user._id
     
     try {
-        const result = await blog.save()
-        response.status(201).json(result)
+        const blogToSave = await blog.save()
+        user.blogs = user.blogs.concat(blogToSave._id)
+        await user.save()
+        response.status(201).json(blogToSave)
     } catch(exception) {
         response.status(400)
         next(exception)
@@ -27,9 +46,20 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 //-------------------------------
-//--------- DELETE ONE ----------
+//------- DELETE ONE BLOG -------
 //-------------------------------
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', async (request, response, next) =>
+{
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log(blog)
+    console.log('Blog user: ', blog.user)
+
+    if (!decodedToken.id || decodedToken.id !== blog.user.toString())
+    {
+        return response.status(401).json({error: 'Authorization failed'})
+    }
+
     try
     {
         const removedBlog = await Blog.findByIdAndDelete(request.params.id)
@@ -45,7 +75,7 @@ blogsRouter.delete('/:id', async (request, response, next) => {
 })
 
 //-------------------------------
-//--------- UPDATE ONE ----------
+//------ UPDATE ONE BLOG --------
 //-------------------------------
 blogsRouter.put('/:id', async (request, response, next) =>
 {
@@ -58,5 +88,6 @@ blogsRouter.put('/:id', async (request, response, next) =>
         next(exception)
     }
 })
+
 
 module.exports = blogsRouter
